requirements.txt
Flask==2.3.2
Flask-WTF==1.1.1
ffmpeg-python==0.2.0
python-dotenv==1.0.0
werkzeug==2.3.7
Pillow==10.2.0
psutil==5.9.8